AvaloniaTest intermittent issue
Sample to check an issue in unit tests where our tests decorated with AvaloniaTest attribute fails intermittently with "System.PlatformNotSupportedException : Operation is not supported on this platform." exception. We saw there was a similar issue(https://github.com/AvaloniaUI/Avalonia/issues/12588) in past which was resolved but we are still facing the issue. We are using Avalonia 11.0.10 version and Nunit testing framework.